Move invoke commands to tools.automation
