+++ tools/automation/generate.py
