+++ tools/automation/vendoring/patches/appdirs.patch
