+++ tools/automation/vendoring/patches/requests.patch
